requirements.txt
streamlit==1.46.1
langchain==0.3.26
langchain-openai==0.3.27
langchain-chroma==0.2.4
python-decouple==3.8